Add files via upload
This files program takes dates from a file and formats them to the x/xx/xxxx format. It takes files inputted by users and checks what dates from that file can be converted to the new format. Once the dates have been changed to the new format, they are written to a new file given the 'parsedDates.txt' file name.